third_party/WebKit/LayoutTests/inspector-protocol/css/css-set-effective-property-value-expected.txt

The test verifies functionality of protocol method CSS.setStyleText.

Dumping matched rules: 
*#inspected* {    regular
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
    }
Dumping inherited rules: 

Running test: testBasicPropertyChange
*#inspected* {    regular
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 101px;
    }

Running test: testChangePropertyInShortHand
*#inspected* {    regular
    margin-left: 10px !important;
}
*#inspected* {    regular
    padding: 10px 20px 101px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 101px;
    padding-left: 40px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
    }

Running test: testChangeImportantProperty
*#inspected* {    regular
    margin-left: 101px !important;
}
*#inspected* {    regular
    padding: 10px 20px 30px 40px;
    padding-top: 50px;
    padding-right: 20px;
    padding-bottom: 30px;
    padding-left: 40px;
}
@media (min-width: 1px)
    *#inspected* {    regular
        padding-left: 5px;
        margin-left: 20px;
        padding-left: 10px;
    }
